Re-enable Compiler extension, test DefaultProvider with both integrations
