+++ Project.toml
@@ -21,7 +21,7 @@ WidthLimitedIO = "b8c1c048-cf81-46c6-9da0-18c1d99e41f2"
 Compiler = "807dbc54-b67e-4c79-8afb-eafe4df6f2e1"
 
 [extensions]
-# CthulhuCompilerExt = "Compiler"
+CthulhuCompilerExt = "Compiler"
 
 [compat]
 Accessors = "0.1.42"
@@ -42,6 +42,7 @@ julia = "1.12"
 [extras]
 CthulhuCompilerExt = "a0401a94-d28a-5d0d-bd4d-a83640b62d95"
 DeepDiffs = "ab62b9b5-e342-54a8-a765-a90f495de1a6"
+Logging = "56ddb016-857b-54e1-b83d-db4d58db5568"
 PerformanceTestTools = "dc46b164-d16f-48ec-a853-60448fc869fe"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Revise = "295af30f-e4ad-537b-8983-00126c2a3abe"
@@ -49,4 +50,4 @@ StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["Test", "DeepDiffs", "PerformanceTestTools", "Random", "Revise", "StaticArrays"]
+test = ["Test", "Compiler", "DeepDiffs", "Logging", "PerformanceTestTools", "Random", "Revise", "StaticArrays"]
